# openapi.json
{
	"openapi": "3.0.0",
	"paths": {
...
		"/api/v1/discounts": {
			"post": {
				"operationId": "DiscountSales_create",
				"parameters": [ ],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateDiscountSaleDto"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DiscountSaleDto"
								}
							}
						}
					}
				},
				"tags": [
					"Discount"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			},
			"get": {
				"operationId": "DiscountSales_findAll",
				"parameters": [ ],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/DiscountSaleDto"
									}
								}
							}
						}
					}
				},
				"tags": [
					"Discount"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			}
		},
		"/api/v1/discounts/{id}": {
			"get": {
				"operationId": "DiscountSales_findOne",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DiscountSaleDto"
								}
							}
						}
					}
				},
				"tags": [
					"Discount"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			},
			"patch": {
				"operationId": "DiscountSales_update",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateDiscountDto"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DiscountSaleDto"
								}
							}
						}
					}
				},
				"tags": [
					"Discount"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			},
			"delete": {
				"operationId": "DiscountSales_remove",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DiscountSaleDto"
								}
							}
						}
					}
				},
				"tags": [
					"Discount"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			}
		},
		"/api/v1/discounts/{id}/products": {
			"put": {
				"operationId": "DiscountSales_addProducts",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DiscountSaleAddProductsDto"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DiscountSaleDto"
								}
							}
						}
					}
				},
				"tags": [
					"Discount"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			}
		},
		"/api/v1/discounts/{id}/categories": {
			"put": {
				"operationId": "DiscountSales_addCategories",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DiscountSaleAddCategoriesDto"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DiscountSaleDto"
								}
							}
						}
					}
				},
				"tags": [
					"Discount"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			}
		},
		"/api/v1/discounts/{id}/variants": {
			"put": {
				"operationId": "DiscountSales_addVariants",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DiscountSaleAddVariantsDto"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DiscountSaleDto"
								}
							}
						}
					}
				},
				"tags": [
					"Discount"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			}
		},
		"/api/v1/discounts/{id}/collections": {
			"put": {
				"operationId": "DiscountSales_addCollections",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DiscountSaleAddCollectionsDto"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DiscountSaleDto"
								}
							}
						}
					}
				},
				"tags": [
					"Discount"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			}
		},
		"/api/v1/discounts/{id}/products/{productId}": {
			"delete": {
				"operationId": "DiscountSales_removeProduct",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					},
					{
						"name": "productId",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DiscountSaleDto"
								}
							}
						}
					}
				},
				"tags": [
					"Discount"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			}
		},
		"/api/v1/discounts/{id}/variants/{variantId}": {
			"delete": {
				"operationId": "DiscountSales_removeVariant",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					},
					{
						"name": "variantId",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DiscountSaleDto"
								}
							}
						}
					}
				},
				"tags": [
					"Discount"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			}
		},
		"/api/v1/discounts/{id}/categories/{categoryId}": {
			"delete": {
				"operationId": "DiscountSales_removeCategory",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					},
					{
						"name": "categoryId",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DiscountSaleDto"
								}
							}
						}
					}
				},
				"tags": [
					"Discount"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			}
		},
		"/api/v1/discounts/{id}/collections/{collectionId}": {
			"delete": {
				"operationId": "DiscountSales_removeCollection",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					},
					{
						"name": "collectionId",
						"required": true,
						"in": "path",
						"schema": {
							"type": "number"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DiscountSaleDto"
								}
							}
						}
					}
				},
				"tags": [
					"Discount"
				],
				"security": [
					{
						"bearer": [ ]
					}
				]
			}
		},
...
	},
	"info": {
		"title": "Nima CMS",
		"description": "The core api for the Nima CMS backend.",
		"version": "0.0.1-alpha",
		"contact": {}
	},
...
	"servers": [
		{
			"url": "http://localhost:3333",
			"description": "Localhost"
		}
	],
	"components": {
		"securitySchemes": {
			"bearer": {
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"type": "http"
			}
		},
		"schemas": {
			"TranslatableDto": {
				"type": "object",
				"properties": {
					"de": {
						"type": "string"
					},
					"it": {
						"type": "string"
					},
					"fr": {
						"type": "string"
					},
					"es": {
						"type": "string"
					},
					"el": {
						"type": "string"
					},
					"en": {
						"type": "string"
					}
				}
			},
			"InputType": {
				"type": "string",
				"enum": [
					"DROPDOWN",
					"MULTISELECT",
					"FILE",
					"REFERENCE",
					"NUMERIC",
					"RICH_TEXT",
					"SWATCH",
					"BOOLEAN",
					"DATE",
					"DATE_TIME"
				]
			},
			"Unit": {
				"type": "string",
				"enum": [
					"CM",
					"M",
					"KM",
					"FT",
					"YD",
					"INCH",
					"SQ_CM",
					"SQ_M",
					"SQ_KM",
					"SQ_FT",
					"SQ_YD",
					"SQ_INCH",
					"CUBIC_MILLIMETER",
					"CUBIC_CENTIMETER",
					"CUBIC_DECIMETER",
					"CUBIC_METER",
					"LITER",
					"CUBIC_FOOT",
					"CUBIC_INCH",
					"CUBIC_YARD",
					"QT",
					"PINT",
					"FL_OZ",
					"ACRE_IN",
					"ACRE_FT",
					"G",
					"LB",
					"OZ",
					"KG",
					"TONNE"
				]
			},
...
			"AttributeValueDto": {
				"type": "object",
				"properties": {
					"id": {
						"type": "number",
						"example": 1
					},
					"name": {
						"example": {
							"en": "Green"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"slug": {
						"type": "string",
						"example": "green"
					},
					"sortOrder": {
						"type": "number",
						"example": 1
					},
					"value": {
						"type": "string",
						"example": "Green"
					},
					"fileUrl": {
						"type": "string"
					},
					"richText": {
						"example": {
							"en": "Green"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"boolean": {
						"type": "boolean"
					},
					"dateTime": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name"
				]
			},
...
			"ProductAttributeDto": {
				"type": "object",
				"properties": {
					"id": {
						"type": "number"
					},
					"name": {
						"example": {
							"en": "Attribute Name"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"slug": {
						"type": "string"
					},
					"metadata": {
						"type": "object",
						"example": {}
					},
					"privateMetadata": {
						"type": "object",
						"example": {}
					},
					"availableInGrid": {
						"type": "boolean",
						"example": true
					},
					"visibleInStorefront": {
						"type": "boolean",
						"example": true
					},
					"filterableInDashboard": {
						"type": "boolean",
						"example": true
					},
					"filterableInStorefront": {
						"type": "boolean",
						"example": true
					},
					"valueRequired": {
						"type": "boolean",
						"example": true
					},
					"storefrontSearchPosition": {
						"type": "number",
						"example": 0
					},
					"inputType": {
						"example": "DROPDOWN",
						"$ref": "#/components/schemas/InputType"
					},
					"unit": {
						"$ref": "#/components/schemas/Unit"
					},
					"values": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/AttributeValueDto"
						}
					}
				},
				"required": [
					"id",
					"name",
					"slug",
					"metadata",
					"privateMetadata",
					"availableInGrid",
					"visibleInStorefront",
					"filterableInDashboard",
					"filterableInStorefront",
					"valueRequired",
					"storefrontSearchPosition",
					"inputType",
					"values"
				]
			},
			"MediaDto": {
				"type": "object",
				"properties": {
					"id": {
						"type": "number",
						"example": 1
					},
					"mimeType": {
						"type": "string",
						"example": "image/png"
					},
					"name": {
						"type": "string",
						"example": "Media Name"
					},
					"slug": {
						"type": "string",
						"example": "media-name"
					},
					"url": {
						"type": "string",
						"example": ""
					},
					"thumbnailUrl": {
						"type": "string",
						"example": "",
						"nullable": true
					},
					"alt": {
						"example": {
							"en": "alt text"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"byteSize": {
						"type": "number",
						"example": 1024
					},
					"created": {
						"type": "string",
						"example": "2022-01-01"
					}
				},
				"required": [
					"id",
					"mimeType",
					"name",
					"slug",
					"url",
					"thumbnailUrl",
					"alt",
					"byteSize",
					"created"
				]
			},
			"SortableMediaDto": {
				"type": "object",
				"properties": {
					"sortOrder": {
						"type": "number",
						"example": 1
					},
					"media": {
						"$ref": "#/components/schemas/MediaDto"
					}
				},
				"required": [
					"sortOrder",
					"media"
				]
			},
			"ProductVariantDto": {
				"type": "object",
				"properties": {
					"id": {
						"type": "number",
						"example": 1
					},
					"sku": {
						"type": "string",
						"example": "product-sku-1"
					},
					"name": {
						"example": {
							"en": "Product Variant Name"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"weight": {
						"type": "number",
						"example": 1
					},
					"metadata": {
						"type": "object",
						"example": {}
					},
					"privateMetadata": {
						"type": "object",
						"example": {}
					},
					"sortOrder": {
						"type": "number",
						"example": 1
					},
					"created": {
						"type": "string",
						"example": "2022-01-01T00:00:00.000+0200"
					},
					"updatedAt": {
						"type": "string",
						"example": "2022-01-01T00:00:00.000+0200"
					},
					"currency": {
						"type": "string",
						"example": "EUR"
					},
					"priceAmount": {
						"type": "number",
						"example": 1.4
					},
					"discountedPrice": {
						"type": "number",
						"example": 1.4
					},
					"stock": {
						"type": "number",
						"example": 10
					},
					"trackInventory": {
						"type": "boolean",
						"example": true
					},
					"attributes": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/ProductAttributeDto"
						}
					},
					"productMedia": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/SortableMediaDto"
						}
					}
				},
				"required": [
					"id",
					"sku",
					"name",
					"metadata",
					"privateMetadata",
					"created",
					"updatedAt",
					"currency",
					"stock",
					"trackInventory",
					"attributes",
					"productMedia"
				]
			},
			"ProductDto": {
				"type": "object",
				"properties": {
					"id": {
						"type": "number",
						"example": 1
					},
					"name": {
						"example": {
							"en": "Product Name"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"description": {
						"example": {
							"en": "Product Description"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"updatedAt": {
						"type": "string",
						"example": ""
					},
					"seoDescription": {
						"type": "string",
						"example": "SEO Description"
					},
					"seoTitle": {
						"type": "string",
						"example": "SEO Title"
					},
					"chargeTaxes": {
						"type": "boolean",
						"example": true
					},
					"weight": {
						"type": "number",
						"example": 12.3
					},
					"metadata": {
						"type": "object",
						"example": {}
					},
					"privateMetadata": {
						"type": "object",
						"example": {}
					},
					"slug": {
						"type": "string",
						"example": "product-name"
					},
					"defaultVariantId": {
						"type": "number"
					},
					"descriptionPlaintext": {
						"type": "string",
						"example": "Product Description"
					},
					"rating": {
						"type": "number",
						"example": 10
					},
					"created": {
						"type": "string",
						"example": "2022-01-01"
					},
					"isPublished": {
						"type": "boolean",
						"example": true
					},
					"isVisibleInListings": {
						"type": "boolean",
						"example": true
					},
					"isAvailableForPurchase": {
						"type": "boolean",
						"example": true
					},
					"currency": {
						"type": "string",
						"example": "EUR"
					},
					"minPrice": {
						"type": "number",
						"example": 12.3
					},
					"productTypeId": {
						"type": "number"
					},
					"categoryId": {
						"type": "number"
					},
					"attributes": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/ProductAttributeDto"
						}
					},
					"productMedia": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/SortableMediaDto"
						}
					},
					"collections": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/CollectionDto"
						}
					},
					"defaultVariant": {
						"$ref": "#/components/schemas/ProductVariantDto"
					}
				},
				"required": [
					"id",
					"name",
					"description",
					"updatedAt",
					"seoDescription",
					"seoTitle",
					"chargeTaxes",
					"weight",
					"metadata",
					"privateMetadata",
					"slug",
					"defaultVariantId",
					"descriptionPlaintext",
					"rating",
					"created",
					"isPublished",
					"isVisibleInListings",
					"isAvailableForPurchase",
					"currency",
					"minPrice",
					"productTypeId",
					"categoryId",
					"attributes",
					"productMedia",
					"collections"
				]
			},
			"CollectionProductsDto": {
				"type": "object",
				"properties": {
					"sortOrder": {
						"type": "number"
					},
					"product": {
						"$ref": "#/components/schemas/ProductDto"
					}
				},
				"required": [
					"sortOrder",
					"product"
				]
			},
			"CollectionDto": {
				"type": "object",
				"properties": {
					"id": {
						"type": "number",
						"example": 1
					},
					"name": {
						"example": {
							"en": "Collection Name"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"slug": {
						"type": "string",
						"example": "collection-name"
					},
					"seoDescription": {
						"example": "SEO Description",
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"seoTitle": {
						"example": "SEO Title",
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"description": {
						"example": {
							"en": "Collection Description"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"metadata": {
						"type": "object",
						"example": {}
					},
					"privateMetadata": {
						"type": "object",
						"example": {}
					},
					"backgroundImage": {
						"$ref": "#/components/schemas/MediaDto"
					},
					"products": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/CollectionProductsDto"
						}
					}
				},
				"required": [
					"id",
					"name",
					"slug",
					"seoDescription",
					"seoTitle",
					"description",
					"metadata",
					"privateMetadata",
					"backgroundImage",
					"products"
				]
			},
...
			"CategoryDto": {
				"type": "object",
				"properties": {
					"id": {
						"type": "number",
						"example": 1
					},
					"privateMetadata": {
						"type": "object",
						"example": {}
					},
					"metadata": {
						"type": "object",
						"example": {}
					},
					"seoTitle": {
						"example": {
							"en": "Category Seo Name"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"seoDescription": {
						"example": {
							"en": "Category Seo Description"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"name": {
						"example": {
							"en": "Category Name"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"description": {
						"example": {
							"en": "Category Description"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"slug": {
						"type": "string",
						"example": "category-name"
					},
					"children": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/CategoryDto"
						}
					},
					"parent": {
						"description": "Only for ancestors query",
						"allOf": [
							{
								"$ref": "#/components/schemas/CategoryDto"
							}
						]
					}
				},
				"required": [
					"id",
					"privateMetadata",
					"metadata",
					"seoTitle",
					"seoDescription",
					"name",
					"description",
					"slug",
					"children"
				]
			},
...
			"DiscountType": {
				"type": "string",
				"enum": [
					"PERCENTAGE",
					"FLAT"
				]
			},
			"CreateDiscountSaleDto": {
				"type": "object",
				"properties": {
					"name": {
						"example": {
							"en": "Discount Sale Name"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"discountType": {
						"$ref": "#/components/schemas/DiscountType"
					},
					"discountValue": {
						"type": "number"
					},
					"endDate": {
						"type": "string",
						"example": "2022-12-31"
					},
					"startDate": {
						"type": "string",
						"example": "2022-01-01"
					},
					"metadata": {
						"type": "object",
						"example": {}
					},
					"privateMetadata": {
						"type": "object",
						"example": {}
					}
				},
				"required": [
					"name",
					"discountType",
					"discountValue",
					"metadata",
					"privateMetadata"
				]
			},
			"DiscountSaleDto": {
				"type": "object",
				"properties": {
					"id": {
						"type": "number",
						"example": 1
					},
					"name": {
						"example": {
							"en": "Discount Sale Name"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"discountType": {
						"$ref": "#/components/schemas/DiscountType"
					},
					"discountValue": {
						"type": "number"
					},
					"endDate": {
						"type": "string",
						"example": "2022-12-31"
					},
					"startDate": {
						"type": "string",
						"example": "2022-01-01"
					},
					"metadata": {
						"type": "object",
						"example": {}
					},
					"privateMetadata": {
						"type": "object",
						"example": {}
					},
					"created": {
						"type": "string",
						"example": "2022-01-01"
					},
					"updatedAt": {
						"type": "string",
						"example": "2022-01-01"
					},
					"categories": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/CategoryDto"
						}
					},
					"products": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/ProductDto"
						}
					},
					"variants": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/ProductVariantDto"
						}
					},
					"collections": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/CollectionDto"
						}
					}
				},
				"required": [
					"id",
					"name",
					"discountType",
					"discountValue",
					"metadata",
					"privateMetadata",
					"created",
					"updatedAt",
					"categories",
					"products",
					"variants",
					"collections"
				]
			},
			"UpdateDiscountDto": {
				"type": "object",
				"properties": {
					"name": {
						"example": {
							"en": "Discount Sale Name"
						},
						"allOf": [
							{
								"$ref": "#/components/schemas/TranslatableDto"
							}
						]
					},
					"discountType": {
						"$ref": "#/components/schemas/DiscountType"
					},
					"discountValue": {
						"type": "number"
					},
					"endDate": {
						"type": "string",
						"example": "2022-12-31"
					},
					"startDate": {
						"type": "string",
						"example": "2022-01-01"
					},
					"metadata": {
						"type": "object",
						"example": {}
					},
					"privateMetadata": {
						"type": "object",
						"example": {}
					}
				}
			},
			"DiscountSaleAddProductsDto": {
				"type": "object",
				"properties": {
					"productIds": {
						"type": "array",
						"items": {
							"type": "number"
						}
					}
				},
				"required": [
					"productIds"
				]
			},
			"DiscountSaleAddCategoriesDto": {
				"type": "object",
				"properties": {
					"categoryIds": {
						"type": "array",
						"items": {
							"type": "number"
						}
					}
				},
				"required": [
					"categoryIds"
				]
			},
			"DiscountSaleAddVariantsDto": {
				"type": "object",
				"properties": {
					"variantIds": {
						"type": "array",
						"items": {
							"type": "number"
						}
					}
				},
				"required": [
					"variantIds"
				]
			},
			"DiscountSaleAddCollectionsDto": {
				"type": "object",
				"properties": {
					"collectionIds": {
						"type": "array",
						"items": {
							"type": "number"
						}
					}
				},
				"required": [
					"collectionIds"
				]
			},
...
		}
	}
}